requirements.txt
fastapi
uvicorn
psycopg2-binary
scikit-learn
pandas
python-dotenv
